+++ migrate/changeset/databases/sqlite.py
@@ -8,7 +8,9 @@ from copy import copy
 
 from sqlalchemy.databases import sqlite as sa_base
 
-from migrate.changeset import ansisql, exceptions, SQLA_06
+from migrate import exceptions
+from migrate.changeset import ansisql, SQLA_06
+
 
 if not SQLA_06:
     SQLiteSchemaGenerator = sa_base.SQLiteSchemaGenerator