+++ migrate/versioning/api.py
@@ -29,7 +29,8 @@ import sys
 import inspect
 import logging
 
-from migrate.versioning import (exceptions, repository, schema, version,
+from migrate import exceptions
+from migrate.versioning import (repository, schema, version,
     script as script_) # command name conflict
 from migrate.versioning.util import catch_known_errors, with_engine