+++ migrate/versioning/pathed.py
@@ -6,7 +6,7 @@ import os
 import shutil
 import logging
 
-from migrate.versioning import exceptions
+from migrate import exceptions
 from migrate.versioning.config import *
 from migrate.versioning.util import KeyedInstance