+++ migrate/versioning/script/py.py
@@ -7,12 +7,12 @@ import logging
 from StringIO import StringIO
 
 import migrate
-from migrate.versioning import exceptions, genmodel, schemadiff
+from migrate import exceptions
+from migrate.versioning import genmodel, schemadiff
 from migrate.versioning.config import operations
 from migrate.versioning.template import Template
 from migrate.versioning.script import base
 from migrate.versioning.util import import_path, load_model, with_engine
-from migrate.changeset.exceptions import *
 
 
 log = logging.getLogger(__name__)
@@ -89,7 +89,7 @@ class PythonScript(base.BaseScript):
         
         :param path: Script location
         :type path: string
-        :raises: :exc:`InvalidScriptError <migrate.versioning.exceptions.InvalidScriptError>`
+        :raises: :exc:`InvalidScriptError <migrate.exceptions.InvalidScriptError>`
         :returns: Python module
         """
         # Try to import and get the upgrade() func
@@ -141,7 +141,7 @@ class PythonScript(base.BaseScript):
             script_func(engine)
         except TypeError:
             warnings.warn("upgrade/downgrade functions must accept engine"
-                " parameter (since version > 0.5.4)", MigrateDeprecationWarning)
+                " parameter (since version > 0.5.4)", exceptions.MigrateDeprecationWarning)
             raise
 
     @property