+++ migrate/versioning/util/__init__.py
@@ -11,11 +11,10 @@ from sqlalchemy import create_engine
 from sqlalchemy.engine import Engine
 from sqlalchemy.pool import StaticPool
 
-from migrate.versioning import exceptions
+from migrate import exceptions
 from migrate.versioning.util.keyedinstance import KeyedInstance
 from migrate.versioning.util.importpath import import_path
 
-from migrate.changeset import *
 
 log = logging.getLogger(__name__)
 
@@ -31,7 +30,7 @@ def load_model(dotted_name):
         if ':' not in dotted_name:
             # backwards compatibility
             warnings.warn('model should be in form of module.model:User '
-                'and not module.model.User', MigrateDeprecationWarning)
+                'and not module.model.User', exceptions.MigrateDeprecationWarning)
             dotted_name = ':'.join(dotted_name.rsplit('.', 1))
         return EntryPoint.parse('x=%s' % dotted_name).load(False)
     else:
@@ -127,7 +126,7 @@ def construct_engine(engine, **opts):
     if echo:
         warnings.warn('echo=True parameter is deprecated, pass '
             'engine_arg_echo=True or engine_dict={"echo": True}',
-            MigrateDeprecationWarning)
+            exceptions.MigrateDeprecationWarning)
         kwargs['echo'] = echo
 
     # parse keyword arguments